--- nodeservice.ts
@@ -1,1458 +1,0 @@
-import { EventEmitter, Injectable } from '@angular/core';
-@Injectable({
-  providedIn: 'root',
-})
-export class NodeService {
-  menuItemClick = new EventEmitter();
-
-  nodes() {
-    return [
-      {
-        id: '1',
-        label: 'Clean Air',
-        icon: 'pi pi-fw pi-plus',
-        target: 'division',
-        expanded: false,
-        command: (event: any) => this.onMenuItemClick(event),
-        items: [
-          {
-            id: '4',
-            label: 'Chengzhou',
-            icon: 'pi pi-fw pi-trash',
-            target: 'plant',
-            expanded: false,
-            command: (event: any) => this.onMenuItemClick(event),
-            items: null,
-          },
-          {
-            id: '3',
-            label: 'Skokie',
-            icon: 'pi pi-fw pi-trash',
-            target: 'plant',
-            expanded: false,
-            command: (event: any) => this.onMenuItemClick(event),
-            items: null,
-          },
-          {
-            id: '1',
-            label: 'Rybnik',
-            icon: 'pi pi-fw pi-trash',
-            target: 'plant',
-            expanded: false,
-            command: (event: any) => this.onMenuItemClick(event),
-            items: [
-              {
-                id: '1',
-                label: 'FF1 -Muffler Box',
-                icon: 'pi pi-fw pi-bookmark',
-                target: 'focusFactory',
-                expanded: false,
-                command: (event: any) => this.onMenuItemClick(event),
-                items: [
-                  {
-                    id: '37',
-                    label: 'FF1Zone09',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: [
-                      {
-                        id: '44',
-                        label: 'Inserter Zone9',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '55',
-                        label: 'Inserter1 Zone9',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '56',
-                        label: 'Inserter2 Zone9',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '57',
-                        label: 'Inserter3 Zone9',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '58',
-                        label: 'Inserter4 Zone9',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                    ],
-                  },
-                  {
-                    id: '34',
-                    label: 'FF1Zone07',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '35',
-                    label: 'FF1Zone08',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '54',
-                    label: 'FF1Zone10',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '55',
-                    label: 'FF1Zone11',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '56',
-                    label: 'FF1Zone12',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '57',
-                    label: 'FF1Zone13',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '58',
-                    label: 'FF1Zone14',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '1',
-                    label: 'FF1Zone01',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: [
-                      {
-                        id: '7',
-                        label: 'Pipe Perforation',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '4',
-                        label: 'Eagle_2',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '17',
-                        label: 'Goliat_1',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '8',
-                        label: 'Inserts',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '5',
-                        label: 'Eagle_5',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '9',
-                        label: 'Cutting Line',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '6',
-                        label: 'Goliat_5',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '3',
-                        label: 'Eagle_1',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '45',
-                        label: 'Eagle_3',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '46',
-                        label: 'Eagle_4',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '47',
-                        label: 'Eagle_6',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '48',
-                        label: 'CuttingLine01',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '49',
-                        label: 'CuttingLine02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '50',
-                        label: 'CuttingLine03',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '51',
-                        label: 'CuttingLine04',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                    ],
-                  },
-                  {
-                    id: '19',
-                    label: 'FF1Zone02',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: [
-                      {
-                        id: '23',
-                        label: 'Inserter',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '20',
-                        label: 'Pipe Perfortation',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '59',
-                        label: 'Eagle01 Zone02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '60',
-                        label: 'Eagle02 Zone02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '61',
-                        label: 'Eagle03 Zone02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '62',
-                        label: 'Goliate02 Zone02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '22',
-                        label: 'Goliate01 Zone02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '63',
-                        label: 'Goliate03 Zone02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '64',
-                        label: 'Goliate04 Zone02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '65',
-                        label: 'Goliate05 Zone02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                    ],
-                  },
-                  {
-                    id: '20',
-                    label: 'FF1Zone03',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: [
-                      {
-                        id: '25',
-                        label: 'PipePerforation',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '26',
-                        label: 'Inserter01',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '27',
-                        label: 'Eagle005',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '28',
-                        label: 'Eagle002',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '29',
-                        label: 'Goliate001',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                    ],
-                  },
-                  {
-                    id: '21',
-                    label: 'FF1Zone04',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: [
-                      {
-                        id: '30',
-                        label: 'Zone04 Inserter',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '31',
-                        label: 'Zone04 Goliate',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '52',
-                        label: 'Zone04 Inster01',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '53',
-                        label: 'Zone04 Inster02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '54',
-                        label: 'Zone04 Inster03',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                    ],
-                  },
-                  {
-                    id: '22',
-                    label: 'FF1Zone05',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '59',
-                    label: 'FF1Zone15',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '60',
-                    label: 'FF1Zone16',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '61',
-                    label: 'FF1Zone17',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '62',
-                    label: 'FF1Zone18',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '63',
-                    label: 'FF1Zone19',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                ],
-              },
-              {
-                id: '2',
-                label: 'FF2 -Cutting Line',
-                icon: 'pi pi-fw pi-bookmark',
-                target: 'focusFactory',
-                expanded: false,
-                command: (event: any) => this.onMenuItemClick(event),
-                items: [
-                  {
-                    id: '23',
-                    label: 'FF2Zone02',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: [
-                      {
-                        id: '36',
-                        label: 'Goliate01',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '37',
-                        label: 'Goliate02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '38',
-                        label: 'Goliate03',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '39',
-                        label: 'Goliate04',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '40',
-                        label: 'Goliate06',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                    ],
-                  },
-                  {
-                    id: '24',
-                    label: 'FF2Zone03',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: [
-                      {
-                        id: '35',
-                        label: 'Cutting Line 04',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                    ],
-                  },
-                  {
-                    id: '3',
-                    label: 'FF2Zone04',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: [
-                      {
-                        id: '41',
-                        label: 'Goliate07',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                    ],
-                  },
-                  {
-                    id: '64',
-                    label: 'FF2Zone05',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '65',
-                    label: 'FF2Zone06',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '66',
-                    label: 'FF2Zone07',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '67',
-                    label: 'FF2Zone08',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '68',
-                    label: 'FF2Zone09',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '69',
-                    label: 'FF2Zone10',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '70',
-                    label: 'FF2Zone11',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '71',
-                    label: 'FF2Zone12',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                ],
-              },
-              {
-                id: '18',
-                label: 'FF6-WeildingLine',
-                icon: 'pi pi-fw pi-bookmark',
-                target: 'focusFactory',
-                expanded: false,
-                command: (event: any) => this.onMenuItemClick(event),
-                items: null,
-              },
-              {
-                id: '19',
-                label: 'FF7-Cleaning',
-                icon: 'pi pi-fw pi-bookmark',
-                target: 'focusFactory',
-                expanded: false,
-                command: (event: any) => this.onMenuItemClick(event),
-                items: [
-                  {
-                    id: '30',
-                    label: 'Zone01-FF07',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '31',
-                    label: 'Zone02-FF07',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '32',
-                    label: 'Zone03-FF07',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '33',
-                    label: 'Zone04-FF07',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                ],
-              },
-              {
-                id: '16',
-                label: 'FF5-Welding Line',
-                icon: 'pi pi-fw pi-bookmark',
-                target: 'focusFactory',
-                expanded: false,
-                command: (event: any) => this.onMenuItemClick(event),
-                items: [
-                  {
-                    id: '79',
-                    label: 'FF5Zone02',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone', 
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '16',
-                    label: 'FF5Zone05',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: [
-                      {
-                        id: '18',
-                        label: 'Eagle 6',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                    ],
-                  },
-                  {
-                    id: '18',
-                    label: 'FF5Zone01',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '80',
-                    label: 'FF5Zone03',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '81',
-                    label: 'FF5Zone04',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                ],
-              },
-              {
-                id: '23',
-                label: 'FF1_2 Muffler Boxes',
-                icon: 'pi pi-fw pi-bookmark',
-                target: 'focusFactory',
-                expanded: false,
-                command: (event: any) => this.onMenuItemClick(event),
-                items: [
-                  {
-                    id: '6',
-                    label: 'FF2Zone01',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: [
-                      {
-                        id: '32',
-                        label: 'Cutting Line01',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '33',
-                        label: 'Cutting Line 02',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '34',
-                        label: 'Cutting Line 03',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '66',
-                        label: 'Cutting coats',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: [
-                          {
-                            id: '170',
-                            label: 'Shift 3',
-                            icon: 'pi pi-fw pi-clock',
-                            target: 'shift',
-                            expanded: false,
-                            command: (event: any) =>
-                              this.onMenuItemClick(event),
-                          },
-                          {
-                            id: '168',
-                            label: 'Shift 1',
-                            icon: 'pi pi-fw pi-clock',
-                            target: 'shift',
-                            expanded: false,
-                            command: (event: any) =>
-                              this.onMenuItemClick(event),
-                          },
-                          {
-                            id: '169',
-                            label: 'Shift 2',
-                            icon: 'pi pi-fw pi-clock',
-                            target: 'shift',
-                            expanded: false,
-                            command: (event: any) =>
-                              this.onMenuItemClick(event),
-                          },
-                        ],
-                      },
-                      {
-                        id: '67',
-                        label: 'Comas',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: [
-                          {
-                            id: '170',
-                            label: 'Shift 3',
-                            icon: 'pi pi-fw pi-clock',
-                            target: 'shift',
-                            expanded: false,
-                            command: (event: any) =>
-                              this.onMenuItemClick(event),
-                          },
-                          {
-                            id: '168',
-                            label: 'Shift 1',
-                            icon: 'pi pi-fw pi-clock',
-                            target: 'shift',
-                            expanded: false,
-                            command: (event: any) =>
-                              this.onMenuItemClick(event),
-                          },
-                          {
-                            id: '169',
-                            label: 'Shift 2',
-                            icon: 'pi pi-fw pi-clock',
-                            target: 'shift',
-                            expanded: false,
-                            command: (event: any) =>
-                              this.onMenuItemClick(event),
-                          },
-                        ],
-                      },
-                      {
-                        id: '68',
-                        label: 'DPP',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: [
-                          {
-                            id: '170',
-                            label: 'Shift 3',
-                            icon: 'pi pi-fw pi-clock',
-                            target: 'shift', 
-                            expanded: false,
-                            command: (event: any) =>
-                              this.onMenuItemClick(event),
-                          },
-                          {
-                            id: '168',
-                            label: 'Shift 1',
-                            icon: 'pi pi-fw pi-clock',
-                            target: 'shift',
-                            expanded: false,
-                            command: (event: any) =>
-                              this.onMenuItemClick(event),
-                          },
-                          {
-                            id: '169',
-                            label: 'Shift 2',
-                            icon: 'pi pi-fw pi-clock',
-                            target: 'shift',
-                            expanded: false,
-                            command: (event: any) =>
-                              this.onMenuItemClick(event),
-                          },
-                        ],
-                      },
-                      {
-                        id: '69',
-                        label: 'E4437',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '70',
-                        label: 'Eagle 1',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                    ],
-                  },
-                ],
-              },
-              {
-                id: '24',
-                label: 'FF3_Pipe Production',
-                icon: 'pi pi-fw pi-bookmark',
-                target: 'focusFactory',
-                expanded: false,
-                command: (event: any) => this.onMenuItemClick(event),
-                items: [
-                  {
-                    id: '44',
-                    label: 'FF3Zone11',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                ],
-              },
-              {
-                id: '3',
-                label: 'FF4_Final Assembly',
-                icon: 'pi pi-fw pi-bookmark',
-                target: 'focusFactory',
-                expanded: false,
-                command: (event: any) => this.onMenuItemClick(event),
-                items: [
-                  {
-                    id: '45',
-                    label: 'FF3Zone12',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '43',
-                    label: 'FF3Zone10',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '46',
-                    label: 'FF3Zone13',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '47',
-                    label: 'FF3Zone14',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '48',
-                    label: 'FF3Zone15',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '49',
-                    label: 'FF3Zone16',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '50',
-                    label: 'FF3Zone17',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '51',
-                    label: 'FF3Zone18',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '52',
-                    label: 'FF3Zone19',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '53',
-                    label: 'FF3Zone20',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '2',
-                    label: 'FF3Zone02',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '14',
-                    label: 'FF3Zone05',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '26',
-                    label: 'FF3Zone04',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '38',
-                    label: 'FF3Zone03',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '39',
-                    label: 'FF3Zone06',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '40',
-                    label: 'FF3Zone07',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '41',
-                    label: 'FF3Zone08',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '42',
-                    label: 'FF3Zone09',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '25',
-                    label: 'FF3Zone01',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '10',
-                    label: 'FF4Zone01',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: [
-                      {
-                        id: '71',
-                        label: 'BMW K4x - manifolds (LM021)',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '72',
-                        label: 'DOP',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                      {
-                        id: '73',
-                        label: 'DOP - Polishing shop',
-                        icon: 'pi pi-fw pi-users',
-                        target: 'line',
-                        expanded: false,
-                        command: (event: any) => this.onMenuItemClick(event),
-                        items: null,
-                      },
-                    ],
-                  },
-                ],
-              },
-              {
-                id: '4',
-                label: 'FF4 - Assembly Line',
-                icon: 'pi pi-fw pi-bookmark',
-                target: 'focusFactory',
-                expanded: false,
-                command: (event: any) => this.onMenuItemClick(event),
-                items: [
-                  {
-                    id: '72',
-                    label: 'FF4Zone04',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '73',
-                    label: 'FF4Zone05',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '27',
-                    label: 'FF4Zone02',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '28',
-                    label: 'FF4Zone03',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '74',
-                    label: 'FF4Zone06',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '75',
-                    label: 'FF4Zone07',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '76',
-                    label: 'FF4Zone08',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '77',
-                    label: 'FF4Zone09',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                  {
-                    id: '78',
-                    label: 'FF4Zone10',
-                    icon: 'pi pi-fw pi-calendar',
-                    target: 'zone',
-                    expanded: false,
-                    command: (event: any) => this.onMenuItemClick(event),
-                    items: null,
-                  },
-                ],
-              },
-            ],
-          },
-          {
-            id: '2',
-            label: 'Edenkoben',
-            icon: 'pi pi-fw pi-trash',
-            target: 'plant',
-            expanded: false,
-            command: (event: any) => this.onMenuItemClick(event),
-            items: null,
-          },
-        ],
-      },
-      {
-        id: '26',
-        label: 'Division Name1',
-        icon: 'pi pi-fw pi-plus',
-        target: 'division',
-        expanded: false,
-        command: (event: any) => this.onMenuItemClick(event),
-        items: [
-          {
-            id: '5',
-            label: 'Port Elizabeth',
-            icon: 'pi pi-fw pi-trash',
-            target: 'plant',
-            expanded: false,
-            command: (event: any) => this.onMenuItemClick(event),
-            items: null,
-          },
-        ],
-      },
-      {
-        id: '2',
-        label: 'Powertrain',
-        icon: 'pi pi-fw pi-plus',
-        target: 'division',
-        expanded: false,
-        command: (event: any) => this.onMenuItemClick(event),
-        items: null,
-      },
-      {
-        id: '4',
-        label: 'Performance Solution',
-        icon: 'pi pi-fw pi-plus',
-        target: 'division',
-        expanded: false,
-        command: (event: any) => this.onMenuItemClick(event),
-        items: null,
-      },
-      {
-        id: '3',
-        label: 'DRiV',
-        icon: 'pi pi-fw pi-plus',
-        target: 'division',
-        expanded: false,
-        command: (event: any) => this.onMenuItemClick(event),
-        items: null,
-      },
-      {
-        id: '52',
-        label: 'Power Drain',
-        icon: 'pi pi-fw pi-plus',
-        target: 'division',
-        expanded: false,
-        command: (event: any) => this.onMenuItemClick(event),
-        items: null,
-      },
-    ];
-  }
-
-  onMenuItemClick(arg0: any) {
-    this.menuItemClick.next(arg0);
-  }
-
-  getNodes() {
-    return Promise.resolve(this.nodes());
-  }
-}
