+++ src/app/@imports/components/meu.filters.const.ts
@@ -1,0 +1,9 @@
+import { NodeContainerComponent } from "../../_modules/hierarchy-node/node-container/node-container.component";
+import { OaeFilterComponent } from "../../_modules/hierarchy-node/oae-filter/oae-filter.component";
+import { RecursiveItemComponent } from "../../_modules/hierarchy-node/recursive-item/recursive-item.component";
+
+const declarations = [
+    OaeFilterComponent,
+    NodeContainerComponent,
+    RecursiveItemComponent
+]
