+++ src/app/_modules/hierarchy-node/node-container/node-container.component.scss
@@ -1,0 +1,23 @@
+
+  div#container_recursive_item {
+    padding: 0.25rem 0;
+    background: #ffffff;
+    color: #495057;
+    box-shadow: 1px 2px 8px 1px rgba(0, 0, 0, 0.3);
+    border-radius: 6px;
+    font-family: inherit;
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
+    font-size: 1rem;
+    font-weight: normal;
+    ul {
+      list-style: none;
+      padding: 0;
+      margin: 0;
+    }
+}
+
+
+.flex-item {
+  font-size: calc(0.8em + 0.11vw);
+}
+
