+++ src/app/_modules/hierarchy-node/node-container/node-container.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { NodeContainerComponent } from './node-container.component';
+
+describe('NodeContainerComponent', () => {
+  let component: NodeContainerComponent;
+  let fixture: ComponentFixture<NodeContainerComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ NodeContainerComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(NodeContainerComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
