+++ src/app/_modules/hierarchy-node/oae-filter/oae-filter.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { OaeFilterComponent } from './oae-filter.component';
+
+describe('OaeFilterComponent', () => {
+  let component: OaeFilterComponent;
+  let fixture: ComponentFixture<OaeFilterComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ OaeFilterComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(OaeFilterComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
