+++ src/app/app.component.html
@@ -1,0 +1,5 @@
+
+<main class="main">
+    <filter-menu></filter-menu>
+</main>
+<router-outlet />
